Added account deletion, github still in progress

+++ app/Http/Controllers/SettingsController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\Hash;
 use Illuminate\Http\RedirectResponse;
 use Illuminate\Support\Facades\Redirect;
 use App\Http\Controllers\Auth\LoginController;
@@ -24,17 +25,15 @@ class SettingsController extends Controller
 
     public function update(Request $request)
     {
-
-        // dd($request->username, $request->email);
         $request->validate([
             'username' => ['required', 'string', 'max:255'],
             'email' => ['nullable', 'string', 'email', 'max:255', 'unique:users,email,' . $request->user()->id],
         ]);
 
         if ($request->user()->provider_id !== null) {
-            $request->validate([
-                'email' => ['nullable'],
-            ]);
+            if ($request->email !== $request->user()->email) {
+                return Redirect::route('settings.edit')->with('status', 'email-not-updated');
+            }
         }
 
         $request->user()->username = $request->username;
@@ -43,24 +42,31 @@ class SettingsController extends Controller
             $request->user()->email = $request->email;
         }
 
-
         $request->user()->save();
 
         return Redirect::route('settings.edit')->with('status', 'username-updated');
     }
 
-    public function destroy(Request $request): RedirectResponse
+    public function destroy(Request $request)
     {
-        $request->validateWithBag('userDeletion', [
-            'password' => ['required', 'current-password'],
+        $request->validate([
+            'password' => ['required', 'string', 'min:8'],
         ]);
+
+        if (!Hash::check($request->password, $request->user()->password)) {
+            return Redirect::route('settings.edit')->with('status', 'password-incorrect');
+        }
 
         $user = $request->user();
 
         Auth::logout();
 
+        $request->session()->invalidate();
+
+        $request->session()->regenerateToken();
+
         $user->delete();
 
-        return $this->loginController->sessionInvalidationAndRedirect($request);
+        return Redirect::route('blogpost.index')->with('status', 'account-deleted');
     }
 }
